+++ documentation/changelog.md
@@ -4,12 +4,12 @@
 - Load Autoback
 - Fade shapes
 - Tooltip/Instructions to use variable in text shape with {{ var_name }}
-- Can't remove colors from shape object
 - rect needs a fill, stroke option
 - responsive broke text entry (no longer aligned)
 - use anchor links to restore page user was last on before leaving
 
 ## Changelog:
+- 2018-08-21 Fixed unable to set or delete shape color in objects.
 - 2018-08-20 Fixed unable to delete shape drawcode.
 - 2018-08-20 Added preview image to Object properties.
 - 2018-08-17 Added preview image and objects using image to Image properties.